KynZo
...

+++ src/Controller/ProjetsController.php
@@ -76,17 +76,19 @@ class ProjetsController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             // on récupere les img uploads
             $images = $form->get('ImgProjet')->getData();
-            // on gere le nom de l'img
-            $fichier = md5(uniqid()) . '.' . $images->guessExtension();
-            // on copie le fichier dans le dossier
-            $images->move(
-                $this->getParameter('images_directory'),
-                $fichier
-            );
-            $nom = $projet->getImgProjet();
-            unlink($this->getParameter('images_directory').'/'.$nom);
-            // stocke le name de l'img dans la BDD
-            $projet->setImgProjet($fichier);
+            if ($images != null) {
+                // on gere le nom de l'img
+                $fichier = md5(uniqid()) . '.' . $images->guessExtension();
+                // on copie le fichier dans le dossier
+                $images->move(
+                    $this->getParameter('images_directory'),
+                    $fichier
+                );
+                $nom = $projet->getImgProjet();
+                unlink($this->getParameter('images_directory').'/'.$nom);
+                // stocke le name de l'img dans la BDD
+                $projet->setImgProjet($fichier);
+            }
             $this->getDoctrine()->getManager()->flush();
 
             $this->addFlash('success', 'Projet modifié avec succès');
